+++ cmd/models.go
@@ -3,7 +3,7 @@
 package main
 
 import (
-	_ "github.com/azure/kaito/presets/falcon"
-	_ "github.com/azure/kaito/presets/llama2"
-	_ "github.com/azure/kaito/presets/llama2chat"
+	_ "github.com/azure/kaito/presets/models/falcon"
+	_ "github.com/azure/kaito/presets/models/llama2"
+	_ "github.com/azure/kaito/presets/models/llama2chat"
 )
